Create nwfsc_combo_survey_maturity.R
An initial pull of maturity data in the NWFSC survey...no analysis, feel free to explore. Note, maturity information will be updated by Melissa Head (NWFSC) (and these data may change). But this is likely the data used in the 2013 sensitivity analysis (which noted odd patterns and the need for more data to inform a maturity curve)