Plotting Updates
Code changed in SST_discards.R. and fishery_landings.R.

In discards, a figure is generated using the GEMM data showing discards grouped by Trawl and Nontrawl by year in mt.

There is also code that creates a histrogram of length comps for discards, but it's odd and I'd like someone else's input.

Additionally, a csv file was exported with specific discard rates by year and fleet.

This file was used in fishery_landings.R to create a stacked bar plot of total landings and discard rates applied to total landings. There is a plot that facets by fleet, and one that aggregates.